+++ src/audible_cli/cmds/cmd_quickstart.py
@@ -3,10 +3,10 @@ import pathlib
 import sys
 
 import click
-from click import echo, prompt, secho
 from tabulate import tabulate
 
 from .. import __version__
+from .._dialog import ask, confirm, say
 from ..config import ConfigFile
 from ..constants import AVAILABLE_MARKETPLACES, CONFIG_FILE, DEFAULT_AUTH_FILE_EXTENSION
 from ..decorators import pass_session
@@ -37,8 +37,8 @@ def ask_user(config: ConfigFile):
     d = {}
     welcome_message = (
         f"\nWelcome to the audible-cli {__version__} quickstart utility.")
-    secho(welcome_message, bold=True)
-    secho(len(welcome_message) * "=", bold=True)
+    say(welcome_message, bold=True)
+    say(len(welcome_message) * "=", bold=True)
 
     intro = """Quickstart will guide you through the process of build a basic
 config, create a first profile and assign an auth file to the profile now.
@@ -51,51 +51,50 @@ an existing auth file when asked about it. Auth files have to be stored in the
 config dir. If the auth file doesn't exists, it will be created. In this case,
 an authentication to the audible server is necessary to register a new device.
 """
-    echo()
-    secho(intro)
+    say()
+    say(intro)
 
     path = config.dirname.absolute()
-    secho("Selected dir to proceed with:", bold=True)
-    echo(path)
+    say("Selected dir to proceed with:", bold=True)
+    say(path)
 
-    echo()
-    echo("Please enter values for the following settings (just press Enter "
+    say()
+    say("Please enter values for the following settings (just press Enter "
          "to accept a default value, if one is given in brackets).")
 
-    echo()
-    d["profile_name"] = prompt(
+    say()
+    d["profile_name"] = ask(
         "Please enter a name for your primary profile",
         default="audible")
 
-    echo()
-    d["country_code"] = prompt(
+    say()
+    d["country_code"] = ask(
         "Enter a country code for the profile",
         show_choices=False,
         type=click.Choice(AVAILABLE_MARKETPLACES)
     )
 
-    echo()
-    d["auth_file"] = prompt(
+    say()
+    d["auth_file"] = ask(
         "Please enter a name for the auth file",
         default=d["profile_name"] + "." + DEFAULT_AUTH_FILE_EXTENSION)
 
     while (path / d["auth_file"]).exists():
-        echo()
-        secho("The auth file already exists in config dir.", bold=True)
-        echo()
+        say()
+        say("The auth file already exists in config dir.", bold=True)
+        say()
 
-        d["use_existing_auth_file"] = click.confirm(
+        d["use_existing_auth_file"] = confirm(
             "Should this file be used for the new profile",
             default=False)
 
         if d["use_existing_auth_file"]:
-            echo()
-            echo("Use existing auth file for new profile.")
+            logger.info("Use existing auth file for new profile.")
 
             return d
 
-        echo()
-        d["auth_file"] = prompt(
+        say()
+        d["auth_file"] = ask(
             "Please enter a new name for the auth file "
             "(or just Enter to exit)",
             default=""
@@ -103,32 +102,32 @@ an authentication to the audible server is necessary to register a new device.
         if not d["auth_file"]:
             sys.exit(1)
 
-    echo()
-    encrypt_file = click.confirm(
+    say()
+    encrypt_file = confirm(
         "Do you want to encrypt the auth file?",
         default=False)
 
     if encrypt_file:
-        echo()
-        d["auth_file_password"] = prompt(
+        say()
+        d["auth_file_password"] = ask(
             "Please enter a password for the auth file",
             confirmation_prompt=True, hide_input=True)
 
-    echo()
-    d["external_login"] = click.confirm(
+    say()
+    d["external_login"] = confirm(
         "Do you want to login with external browser?",
         default=False)
     d["audible_username"] = None
     d["audible_password"] = None
 
-    echo()
-    d["with_username"] = click.confirm(
+    say()
+    d["with_username"] = confirm(
         "Do you want to login with a pre-amazon Audible account?",
         default=False)
 
     if not d["external_login"]:
-        d["audible_username"] = prompt("Please enter your amazon username")
-        d["audible_password"] = prompt(
+        d["audible_username"] = ask("Please enter your amazon username")
+        d["audible_password"] = ask(
             "Please enter your amazon password",
             hide_input=True, confirmation_prompt=True
         )
@@ -150,9 +149,9 @@ def cli(session):
 
     d = ask_user(config)
 
-    echo()
-    echo(tabulate_summary(d))
-    click.confirm("Do you want to continue?", abort=True)
+    say()
+    say(tabulate_summary(d))
+    confirm("Do you want to continue?", abort=True)
 
     if "use_existing_auth_file" not in d:
         build_auth_file(